backend/requirements.txt
aiosmtplib==3.0.1
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.4
asttokens==2.4.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Deprecated==1.2.13
dill==0.3.8
executing==2.0.1
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
icecream==2.1.3
idna==3.7
iniconfig==2.0.0
invoke==2.2.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
mando==0.7.1
MarkupSafe==2.1.5
mccabe==0.7.0
mysql-connector-python==9.0.0
omegaconf==2.3.0
packaging==24.1
phonenumbers==8.12.48
pkce==1.0.3
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.10.4
Pygments==2.18.0
PyJWT==2.8.0
pylint==3.2.6
pytest==8.3.2
python-box==7.2.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
radon==6.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
supertokens-python==0.22.1
timeout-decorator==0.5.0
tldextract==3.1.0
tomlkit==0.13.2
twilio==7.9.1
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xenon==0.9.1
